Added a base class to inherit methods that allow payload loading and JSON return.
