+++ src/main/java/com/google/sps/servlets/DeleteAll.java
@@ -3,13 +3,12 @@ package com.google.sps.servlets;
 import com.google.sps.data.DatastoreService;
 import java.io.IOException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /* Deletes all Users stored in Datastore, used for debugging. */
 @WebServlet("/api/delete-all")
-public class DeleteAll extends HttpServlet {
+public class DeleteAll extends ServletTemplate {
 
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -17,11 +16,11 @@ public class DeleteAll extends HttpServlet {
             try{
                 DatastoreService datastoreService = new DatastoreService();
                 datastoreService.deleteAll();
-                response.setContentType("application/json");
-                response.getWriter().println("{\"success\":\"All accounts deleted.\"}");
+                sendJSONResponse(response,"{\"success\":\"All accounts deleted.\"}");
+                return;
             } catch(Exception e){
-                response.setContentType("application/json");
-                response.getWriter().println("{\"error\":\"Something went wrong.\"}");
+                String message = "{\"error\":\"" + e + "\"}" ;
+                sendJSONResponse(response,message);
             }
             
     }
