+++ src/main/java/com/google/sps/servlets/Register.java
@@ -4,62 +4,15 @@ import com.google.gson.Gson;
 import com.google.sps.data.DatastoreService;
 import com.google.sps.data.User;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.PrintWriter;
 
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /* Handles the register page. */
 @WebServlet("/api/register")
-public class Register extends HttpServlet {
-
-    public void sendJSONResponse(HttpServletResponse response, String message) throws IOException{
-        response.setContentType("application/json");
-        PrintWriter out = response.getWriter();
-        out.print(message);
-        out.flush();
-
-    }
-
-    public static String getBody(HttpServletRequest request) throws IOException {
-
-        String body = null;
-        StringBuilder stringBuilder = new StringBuilder();
-        BufferedReader bufferedReader = null;
-    
-        try {
-            InputStream inputStream = request.getInputStream();
-            if (inputStream != null) {
-                bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-                char[] charBuffer = new char[128];
-                int bytesRead = -1;
-                while ((bytesRead = bufferedReader.read(charBuffer)) > 0) {
-                    stringBuilder.append(charBuffer, 0, bytesRead);
-                }
-            } else {
-                stringBuilder.append("");
-            }
-        } catch (IOException ex) {
-            throw ex;
-        } finally {
-            if (bufferedReader != null) {
-                try {
-                    bufferedReader.close();
-                } catch (IOException ex) {
-                    throw ex;
-                }
-            }
-        }
-    
-        body = stringBuilder.toString();
-        return body;
-    }
+public class Register extends ServletTemplate  {
 
     @Override
     // Using the html form, register a new user if the username and email are not already taken.
